backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
SQLAlchemy<2.0.0
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==0.21.0
flask-cors==4.0.0
